Fix compiler error
1. Delete unnecessary code
2. modify Maddress to uint64_t

Issue: https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/IC5U5Z

+++ common_interfaces/heap/heap_allocator.h
@@ -23,7 +23,7 @@ namespace panda {
 class HeapAllocator {
 public:
     using Address32 = uint32_t;
-    using Address = uintptr_t;
+    using Address = uint64_t;
 
     static Address Allocate(size_t size);
     static Address32 Allocate32(size_t size);
